added delete action on updating history fragments

+++ app/src/main/java/com/romandevyatov/bestfinance/data/dao/IncomeHistoryDao.kt
@@ -24,7 +24,7 @@ interface IncomeHistoryDao {
     suspend fun deleteAll()
 
     @Query("DELETE FROM income_history WHERE id = :id")
-    suspend fun deleteById(id: Int)
+    suspend fun deleteById(id: Long)
 
     @Transaction
     @Query("SELECT * FROM income_history")
